pyproject.toml
[tool.poetry]
name = "translationapiplay"
version = "0.1.0"
description = ""
authors = ["Kamaal Farah <kamaal.f1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.98.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
google-cloud-translate = "^3.11.1"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ruff = "^0.0.275"
pre-commit = "^3.3.3"
google-cloud-api-keys = "^0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
